FIX: Scan files when a single server is online (#42)
* FIX: Scan files when a single server is online

This commits makes sure that even when a single server is online,
the plugin is available. This commit also makes sure that a random
and online server is chosen when using with_session without any
specified socket. It used to use the first one.

* FIX: Ensure used sockets are not closed

* FIX: Add a 3 second timeout

* Fix lint

* Refactor ClamAV service pool

* Redistribute responsibilities.

This change builds upon Bianca's idea of create the `ClamAVService` class and reorganized the code based on the following rationale:

`ClamAV` is what glues everything together. It knows how to talk to the S3 store, the plugin store and the `ClamAVService` to display the version, check availability, and scan files. It gets the raw response from ClamAV and translates to a different format which the rest of the plugin relies on.
`ClamAVService` defines an interface of which operations are possible (`online?`, `scan_file`, `version`) and takes care of the socket and session managemente, as well as writting/reading from the socket.
`ClamAVServicePool` resolves the SRV record and instantiates existing services. It tells them how to open a connection.

While working on this, I tried to remove the latter and do everything using only the first two, but I decided to keep it in the end because this design is handy for testing. We rely on the `FakeTCPSocket` for spying on the communication and returning specific responses, and this design let us use without relying on a mocking framework, just dependency injection.

---------

Co-authored-by: Roman Rizzi <rizziromanalejandro@gmail.com>